requirements.txt
numpy~=1.16.2
setuptools~=46.0.0
pyzmq~=18.0.1